refactor big _after_instrument.h in multiple headers
