+++ include/src_tracer/_after_instrument.h
@@ -1,627 +1,53 @@
-// src_tracer.h
+/*
+   src_tracer/_after_instrument.h
+   automatically included after running instrumenter.py
+*/
 
-// This file comes without any standard library include.
-// Reason: It works on pre-processed files.
-// A "#include <stdbool.h>" would result in double included files!
-
-// editable constant definitions
-#ifndef TRACE_BUF_SIZE
-#define TRACE_BUF_SIZE 4096
-#endif
-#ifndef TRACE_USE_SYSCALLS
-  // if TRACE_USE_POSIX_WRITE is not set we use the syscall directly
-  #ifndef TRACE_USE_POSIX_WRITE
-  #define TRACE_USE_POSIX_WRITE
-  #endif
-#endif
-#ifndef EFFICIENT_TEXT_TRACE
-// well it's not for efficiency, more for debugging
-#endif
-// assumes short is 16 bit
-// assumes char is 8 bit
-
-// other constants
-#define PROP_FALSE          0
-#define PROP_TRUE           1
-#define PROP_DEFAULT_VALUE  2
-
-
-// and here comes the rest...
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-// bool is available in C++ but not in C without (see above) include stdbool
-#ifndef __cplusplus
-#ifndef bool
-#define bool _Bool
+// This file comes without any standard library include.
+// Reason: It works on pre-processed files.
+// A "#include <stdbool.h>" would result in double included files!
+#include "src_tracer/_stdinc_replace.h"
+
+// assumes char is 8 bit!
+#include "src_tracer/constants.h"
+
+#include "src_tracer/trace_elem.h"
+#include "src_tracer/trace_buf.h"
+
+// and here comes the rest...
+#ifdef _TRACE_MODE
+    #include "src_tracer/trace_mode.h"
 #endif
+#ifdef _TRACE_MODE_TEXT
+    #include "src_tracer/trace_mode_text.h"
 #endif
-
-extern void _trace_write(const void *buf);
-#ifndef EFFICIENT_TEXT_TRACE
-extern void _trace_write_text(const void *buf, unsigned long count);
-#endif
-
-extern void _trace_open(const char *fname);
-extern void _trace_close(void);
-extern void _trace_before_fork(void);
-extern int _trace_after_fork(int pid);
-
-extern unsigned long long int _trace_setjmp_idx;
-
-#ifndef BYTE_TRACE
-#define _TRACE_IE_BYTE_INIT         0b11111110
-extern unsigned char _trace_ie_byte;
-#endif
-
-extern unsigned char _trace_buf[TRACE_BUF_SIZE];
-extern int _trace_buf_pos;
-
-#define _TRACE_TEST_IE              0b10000000
- #define _TRACE_SET_IE              0b10000000
-
-#define _TRACE_TEST_OTHER           0b11110000
- #define _TRACE_SET_FUNC_4          0b00000000
- #define _TRACE_SET_FUNC_12         0b00010000
- #define _TRACE_SET_FUNC_20         0b00100000
- #define _TRACE_SET_FUNC_28         0b00110000
- #define _TRACE_SET_ELEM_AO         0b01000000
- #define _TRACE_SET_ELEM_PZ         0b01010000
- #define _TRACE_SET_ELEM2_ao        0b01100000
- #define _TRACE_SET_ELEM2_pz        0b01110000
-
-#define _TRACE_TEST_LEN             0b00000111
- #define _TRACE_SET_LEN_0           0b00000000
- #define _TRACE_SET_LEN_8           0b00000001
- #define _TRACE_SET_LEN_16          0b00000010
- #define _TRACE_SET_LEN_32          0b00000011
- #define _TRACE_SET_LEN_64          0b00000100
- #define _TRACE_SET_LEN_PREFIX_res  0b00000101
- #define _TRACE_SET_LEN_STRING_res  0b00000110
- #define _TRACE_SET_LEN_BYTECOUNT   0b00000111
-
-#define _TRACE_TEST_IS_ELEM         0b11100000
- #define _TRACE_SET_IS_ELEM         0b01000000
-
-#define _TRACE_TEST_ELEM            0b11111111
- #define _TRACE_SET_END             'E'
- #define _TRACE_SET_RETURN          'R'
- #define _TRACE_SET_RETURN_TAIL     'S'
- #define _TRACE_SET_FUNC_ANON       'A'
- #define _TRACE_SET_TRY             'T'
- #define _TRACE_SET_UNTRY           'U'
- #define _TRACE_SET_PAUSE           'P'
-/* FUNC_32 always comes with 32 bit function number */
- #define _TRACE_SET_FUNC_32         'C'
-/* 'I' and 'O' could be used instead of
- * _TRACE_IE_BYTE_INIT for faster trace writing */
- #define _TRACE_SET_IF              'I'
- #define _TRACE_SET_ELSE            'O'
-/* 'F' is reserved, use _TRACE_SET_FORK */
- #define _TRACE_SET_FORK_reserved   'F'
-/* 'J' is reserved, use _TRACE_SET_CATCH */
- #define _TRACE_SET_CATCH_reserved  'J'
-/* 'D' is reserved, use _TRACE_SET_DATA */
- #define _TRACE_SET_DATA_reserved   'D'
-
-/* 'X' to '_' are reserved */
-#define _TRACE_TEST_ELEM_reserved   0b11111000
- #define _TRACE_SET_ELEM_reserved   0b01011000
-
-#define _TRACE_TEST_IS_ELEM2        0b11100000
- #define _TRACE_SET_IS_ELEM2        0b01100000
-
-#define _TRACE_TEST_ELEM2           0b11111000
- #define _TRACE_SET_FORK            0b01100000
- #define _TRACE_SET_CATCH           0b01101000
- #define _TRACE_SET_DATA            0b01110000
- #define _TRACE_SET_ELEM2_reserved  0b01111000
-
-#define likely(x)       __builtin_expect((x),1)
-#define unlikely(x)     __builtin_expect((x),0)
-
-#define _TRACE_PUT(c) ;{ \
-    _trace_buf[_trace_buf_pos] = (c); \
-    _trace_buf_pos += 1; \
-    if (_trace_buf_pos == TRACE_BUF_SIZE) { \
-        _trace_write(_trace_buf); \
-        _trace_buf_pos = 0; \
-    } \
-}
-
-#ifdef EFFICIENT_TEXT_TRACE
-#define _TRACE_PUT_TEXT     _TRACE_PUT
-#else
-#define _TRACE_PUT_TEXT(c) ;{ \
-    unsigned char buf[1] = { (c) }; \
-    _trace_write_text(buf, 1); \
-}
-#endif
-
-#ifdef BYTE_TRACE
-
-#define _TRACE_IF()     _TRACE_PUT('I')
-#define _TRACE_ELSE()   _TRACE_PUT('O')
-#define _TRACE_IE_FINISH
-#define _TRACE_IE(if_true) \
-    if (if_true) { \
-        _TRACE_IF(); \
-    } else { \
-        _TRACE_ELSE(); \
-    }
-
-#define _TRACE_CASE(num, bit_cnt) { \
-    for (int i = bit_cnt-1; i >= 0; i--) { \
-        _TRACE_IE(num & (1 << i)); \
-    } \
-}
-#else
-
-// will get optimized as rotate instruction
-#define rotate_8(x, n) \
-    ((x << n) | (x >> (8 - n)))
-
-#define _TRACE_IE_PREPARE_NEXT \
-    if (_trace_ie_byte < 0b11000000) { \
-        _TRACE_PUT(_trace_ie_byte); \
-        _trace_ie_byte = _TRACE_IE_BYTE_INIT; \
-    }
-
-#define _TRACE_IE(if_true) { \
-    _trace_ie_byte <<= 1; \
-    _trace_ie_byte |= (bool)(if_true); \
-    _TRACE_IE_PREPARE_NEXT \
-}
-#define _TRACE_IF() { \
-    _trace_ie_byte = rotate_8(_trace_ie_byte, 1); \
-    _TRACE_IE_PREPARE_NEXT \
-}
-#define _TRACE_ELSE() { \
-    _trace_ie_byte <<= 1; \
-    _TRACE_IE_PREPARE_NEXT \
-}
-#define _TRACE_IE_FINISH \
-    if (_trace_ie_byte != _TRACE_IE_BYTE_INIT) { \
-        _TRACE_PUT(_trace_ie_byte); \
-        _trace_ie_byte = _TRACE_IE_BYTE_INIT; \
-    }
-
-#define _TRACE_CASE(num, bit_cnt) { \
-    _TRACE_IE_FINISH \
-    int bit_cnt_left = bit_cnt; \
-    const int num_left = num; \
-    while (bit_cnt_left >= 6) { \
-        bit_cnt_left -= 6; \
-        _TRACE_PUT(((num_left >> bit_cnt_left) & 0b00111111) | 0b10000000); \
-    } \
-    _trace_ie_byte = (num_left & ~(0b11111111 << bit_cnt_left) & 0b00111111) - (2 << bit_cnt_left); \
-}
-
-#endif // BYTE_TRACE
-
-// functions numbers are now big endian for better conversion
-#define _TRACE_FUNC_CORE(num) { \
-    if ((num) == 0) { \
-        _TRACE_PUT(_TRACE_SET_FUNC_ANON); \
-    } else if ((num) == ((num) & 0xf)) { \
-        _TRACE_PUT(_TRACE_SET_FUNC_4  | (((num) >> 0) & 0xff)); \
-    } else if ((num) == ((num) & 0xfff)) { \
-        _TRACE_PUT(_TRACE_SET_FUNC_12 | (((num) >> 8) & 0xff)); \
-        _TRACE_PUT(((num) >> 0) & 0xff); \
-    } else if ((num) == ((num) & 0xfffff)) { \
-        _TRACE_PUT(_TRACE_SET_FUNC_20 | (((num) >> 16) & 0xff)); \
-        _TRACE_PUT(((num) >> 8) & 0xff); \
-        _TRACE_PUT(((num) >> 0) & 0xff); \
-    } else if ((num) == ((num) & 0xfffffff)) { \
-        _TRACE_PUT(_TRACE_SET_FUNC_28 | (((num) >> 24) & 0xff)); \
-        _TRACE_PUT(((num) >> 16) & 0xff); \
-        _TRACE_PUT(((num) >> 8) & 0xff); \
-        _TRACE_PUT(((num) >> 0) & 0xff); \
-    } else { \
-        _TRACE_PUT(_TRACE_SET_FUNC_32); \
-        _TRACE_PUT(((num) >> 24) & 0xff); \
-        _TRACE_PUT(((num) >> 16) & 0xff); \
-        _TRACE_PUT(((num) >> 8) & 0xff); \
-        _TRACE_PUT(((num) >> 0) & 0xff); \
-    } \
-}
-
-#ifdef _TRACE_POINTER_CALLS_ONLY
-
-#define _TRACE_FUNC(num) \
-    _TRACE_POINTER_CALL_INIT \
-    _TRACE_IE_FINISH \
-    if (_TRACE_CALL_CHECK) { \
-        _TRACE_FUNC_CORE(num) \
-        _TRACE_POINTER_CALL_RESET \
-    } \
-
-#define _TRACE_STATIC_FUNC(num) \
-    _TRACE_POINTER_CALL_INIT \
-    if (_TRACE_CALL_CHECK) { \
-        _TRACE_IE_FINISH \
-        _TRACE_FUNC_CORE(num) \
-        _TRACE_POINTER_CALL_RESET \
-    }
-
-#else
-#define _TRACE_FUNC(num) \
-    _TRACE_IE_FINISH \
-    _TRACE_FUNC_CORE(num)
-#define _TRACE_STATIC_FUNC(num) \
-    _TRACE_FUNC(num)
-#endif
-
-#define _TRACE_NUM_0(type, num) { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(type | _TRACE_SET_LEN_0); \
-}
-
-#define _TRACE_NUM_8(type, num) { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(type | _TRACE_SET_LEN_8); \
-    _TRACE_PUT(((num) >> 0) & 0xff); \
-}
-
-#define _TRACE_NUM_16(type, num) { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(type | _TRACE_SET_LEN_16); \
-    _TRACE_PUT(((num) >> 0) & 0xff); \
-    _TRACE_PUT(((num) >> 8) & 0xff); \
-}
-
-#define _TRACE_NUM_32(type, num) { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(type | _TRACE_SET_LEN_32); \
-    _TRACE_PUT(((num) >> 0) & 0xff); \
-    _TRACE_PUT(((num) >> 8) & 0xff); \
-    _TRACE_PUT(((num) >> 16) & 0xff); \
-    _TRACE_PUT(((num) >> 24) & 0xff); \
-}
-
-#define _TRACE_NUM_64(type, num) { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(type | _TRACE_SET_LEN_64); \
-    _TRACE_PUT(((num) >> 0) & 0xff); \
-    _TRACE_PUT(((num) >> 8) & 0xff); \
-    _TRACE_PUT(((num) >> 16) & 0xff); \
-    _TRACE_PUT(((num) >> 24) & 0xff); \
-    _TRACE_PUT(((num) >> 32) & 0xff); \
-    _TRACE_PUT(((num) >> 40) & 0xff); \
-    _TRACE_PUT(((num) >> 48) & 0xff); \
-    _TRACE_PUT(((num) >> 56) & 0xff); \
-}
-
-// data should be native endian (here little endian)
-#define _TRACE_NUM(type, num) { \
-    unsigned long long _trace_n = (num); \
-    if (_trace_n == 0) { \
-        _TRACE_NUM_0(type, _trace_n); \
-    } else if (_trace_n == (_trace_n & 0xff)) { \
-        _TRACE_NUM_8(type, _trace_n); \
-    } else if (_trace_n == (_trace_n & 0xffff)) { \
-        _TRACE_NUM_16(type, _trace_n); \
-    } else if (_trace_n == (_trace_n & 0xffffffff)) { \
-        _TRACE_NUM_32(type, _trace_n); \
-    } else { \
-        _TRACE_NUM_64(type, _trace_n); \
-    } \
-}
-
-#define NIBBLE_TO_HEX_(n)   (((n) >= 0xa) ? (n) - 0xa + 'a' : (n) + '0')
-#define NIBBLE_TO_HEX(n,i)  NIBBLE_TO_HEX_(((n) >> ((i)*4)) & 0xf)
-
-// Shift twice, otherwise we might run into undefined behavior!
-#define NIBBLE_COUNT(n,c)   (((n) >> (c)*3 >> (c)) != 0)
-
-#ifdef EFFICIENT_TEXT_TRACE
-#define _TRACE_NUM_TEXT(type, num) ;{ \
-    int count; \
-    _TRACE_PUT(type); \
-    for (count = 0; NIBBLE_COUNT((num), count); count++) {}  \
-    for (int i = count-1; i >= 0; i--) { \
-        _TRACE_PUT(NIBBLE_TO_HEX((num), i)); \
-    } \
-}
-#else
-#define _TRACE_NUM_TEXT(type, num) ;{ \
-    unsigned char buf[18]; \
-    int count; \
-    buf[0] = (type); \
-    for (count = 0; NIBBLE_COUNT((num), count); count++) {}  \
-    for (int i = 0; i < count; i++) { \
-        buf[count-i] = NIBBLE_TO_HEX((num), i); \
-    } \
-    _trace_write_text(buf, 1+count); \
-}
-#endif
-
-#define _TRACE_CASE_TEXT(num, bit_cnt) ; \
-    for (int i = bit_cnt-1; i >= 0; i--) { \
-        if (num & (1 << i)) { \
-            _TRACE_PUT_TEXT('I'); \
-        } else { \
-            _TRACE_PUT_TEXT('O'); \
-        } \
-    }
-
-#define _TRACE_RETURN(type) \
-    if (_TRACE_RETURN_CHECK) { \
-        _TRACE_IE_FINISH \
-        _TRACE_PUT(type); \
-    }
-
-#define _TRACE_END() { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(_TRACE_SET_END); \
-}
-
-#define _TRACE_TRY() { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(_TRACE_SET_TRY); \
-}
-
-#define _TRACE_CATCH(cur_idx) { \
-    _TRACE_NUM_16(_TRACE_SET_CATCH, _trace_setjmp_idx - (cur_idx)); \
-    /* _trace_setjmp_idx = cur_idx; */ \
-}
-
-#if 0
-#define _TRACE_TRY_END() { \
-    _TRACE_IE_FINISH \
-    _TRACE_PUT(_TRACE_SET_UNTRY); \
-    _trace_setjmp_idx --; \
-}
-#else
-#define _TRACE_TRY_END() /* nothing here */
-#endif
-
-#define _TRACE_SETJMP(setjmp_stmt) ({ \
-    _trace_setjmp_idx ++; \
-    int cur_setjmp_idx = _trace_setjmp_idx; \
-    _TRACE_TRY(); \
-    int setjmp_res = setjmp_stmt; \
-    if (setjmp_res != 0) { \
-        _TRACE_CATCH(cur_setjmp_idx); \
-    } \
-    setjmp_res; \
-})
-
-// same as the macro version
-// but returns num
-// can be used inside switch conditions
-static inline __attribute__((always_inline)) long long int _trace_num(char type, long long int num) {
-    _TRACE_NUM(type, num);
-    return num;
-}
-
-static inline __attribute__((always_inline)) long long int _trace_num_text(char type, long long int num) {
-    _TRACE_NUM_TEXT(type, num);
-    return num;
-}
-
-static inline __attribute__((always_inline)) bool _trace_condition(bool cond) {
-    _TRACE_IE(cond);
-    return cond;
-}
-
-static inline __attribute__((always_inline)) bool _text_trace_condition(bool cond) {
-    if (cond) {
-        _TRACE_PUT_TEXT('I');
-    } else {
-        _TRACE_PUT_TEXT('O');
-    }
-    return cond;
-}
 
 // for retracing
-
-extern volatile char _retrace_letter;
-extern volatile long long int _retrace_num;
-extern void _retrace_breakpoint(void);
-
-extern volatile int _retrace_fork_count;
-
-#define _RETRACE_ELEM(type, num) ;{ \
-    _retrace_letter = (type); \
-    _retrace_num = (num); \
-    _retrace_breakpoint(); \
-}
-
-#define _RETRACE_FUNC(num) \
-    _TRACE_POINTER_CALL_INIT \
-    if (_TRACE_CALL_CHECK) { \
-        _RETRACE_ELEM('C', num); \
-        _TRACE_POINTER_CALL_RESET; \
-    }
-
-#define _RETRACE_RETURN(type) \
-    if (_TRACE_RETURN_CHECK) { \
-        _RETRACE_ELEM(type, 0); \
-    }
-
-#define _RETRACE_IF() \
-    _RETRACE_ELEM('I', 0)
-
-#define _RETRACE_ELSE() \
-    _RETRACE_ELEM('O', 0)
-
-#define _RETRACE_END() \
-    _RETRACE_ELEM('E', 0)
-
-#define _RETRACE_TRY() \
-    _RETRACE_ELEM('T', 0)
-
-#define _RETRACE_CATCH(cur_idx) { \
-    _RETRACE_ELEM('J', _trace_setjmp_idx - (cur_idx)); \
-    /* _trace_setjmp_idx = cur_idx; */ \
-}
-
-#if 0
-#define _RETRACE_TRY_END() { \
-    _RETRACE_ELEM('U', 0) \
-    _trace_setjmp_idx --; \
-}
-#else
-#define _RETRACE_TRY_END() /* nothing here */
+#ifdef _RETRACE_MODE
+    #include "src_tracer/retrace_mode.h"
+#endif
+#ifdef _CMBC_MODE
+    #include "src_tracer/cbmc_mode.h"
 #endif
 
-#define _RETRACE_SETJMP(setjmp_stmt) ({ \
-    _trace_setjmp_idx ++; \
-    int cur_setjmp_idx = _trace_setjmp_idx; \
-    _RETRACE_TRY(); \
-    int setjmp_res = setjmp_stmt; \
-    if (setjmp_res != 0) { \
-        _RETRACE_CATCH(cur_setjmp_idx); \
-    } \
-    setjmp_res; \
-})
+// for combined tracing and retracing
+#if defined _TRACE_MODE && defined _RETRACE_MODE
+    #include "src_tracer/combined_mode.h"
+#endif
 
-static inline __attribute__((always_inline)) long long int _retrace_elem(char type, long long int num) {
-    _RETRACE_ELEM(type, num);
-    return num;
-}
-
-static inline __attribute__((always_inline)) bool _retrace_condition(bool cond) {
-    if (cond) {
-        _retrace_letter = 'I';
-    } else {
-        _retrace_letter = 'O';
-    }
-    _retrace_num = 0;
-    _retrace_breakpoint();
-    return cond;
-}
-
-static inline __attribute__((always_inline)) int _retrace_after_fork(int fork_val) {
-    if (fork_val != 0) {
-        _retrace_letter = 'I';
-    } else {
-        _retrace_letter = 'O';
-    }
-    _retrace_num = 0;
-    _retrace_breakpoint();
-    return fork_val;
-}
-
-#define _RETRACE_CASE(num, bit_cnt) { \
-    for (int i = bit_cnt-1; i >= 0; i--) { \
-        _retrace_condition(num & (1 << i)); \
-    } \
-}
-
-// for both tracing and retracing
-extern volatile bool _is_retrace_mode;
-
-#define _IS_RETRACE(a,b)    ; \
-    if (_is_retrace_mode) { \
-        a; \
-    } else { \
-        b; \
-    }
-
-static inline __attribute__((always_inline)) bool _is_retrace_condition(bool cond) {
-    if (cond) {
-        _IS_RETRACE(_RETRACE_IF(), _TRACE_IF());
-    } else {
-        _IS_RETRACE(_RETRACE_ELSE(), _TRACE_ELSE());
-    }
-    return cond;
-}
-
-/* This can be used for switch:
- *    switch(        num ) { ... }
- * Annotated:
- *    switch(_SWITCH(num)) { ... }
- * The makro _SWITCH might translate to _is_retrace_switch.
- */
-static inline __attribute__((always_inline)) long long int _is_retrace_switch(long long int num) {
-    _IS_RETRACE(_RETRACE_ELEM('D', num),
-                _TRACE_NUM(_TRACE_SET_DATA, num)
-    )
-    return num;
-}
 
 #define _RETRO(normal, retro) \
     _RETRO_SKIP(normal) \
     _RETRO_ONLY(retro)
 
-/* Pointer calls */
-#ifdef _TRACE_POINTER_CALLS_ONLY
-
-extern bool _trace_pointer_call;
-
-#define _TRACE_POINTER_CALL_INIT \
-    bool _trace_local_return __attribute__((unused)) = 0;
-
-#define _TRACE_POINTER_CALL_SET \
-    _trace_pointer_call = 1;
-
-#define _TRACE_POINTER_CALL_RESET \
-    _trace_pointer_call = 0; \
-    _trace_local_return = 1;
-
-#define _TRACE_CALL_CHECK _trace_pointer_call
-#define _TRACE_RETURN_CHECK _trace_local_return
-
-#define _TRACE_POINTER_CALL(call) ({ \
-    _TRACE_POINTER_CALL_SET; \
-    call; \
-})
-#define _TRACE_POINTER_CALL_AFTER(type, call) ({ \
-    type _trace_call_tmp = call; \
-    _TRACE_POINTER_CALL_SET; \
-    _trace_call_tmp; \
-})
-
-#else
-#define _TRACE_POINTER_CALL_INIT /* nothing here */
-#define _TRACE_CALL_CHECK 1
-#define _TRACE_POINTER_CALL_SET /* nothing here */
-#define _TRACE_POINTER_CALL_RESET /* nothing here */
-#define _TRACE_POINTER_CALL(call) call
-#define _TRACE_POINTER_CALL_AFTER(type, call) call
-#define _TRACE_RETURN_CHECK 1
-#endif
-
-/* return after */
-#define _FUNC_RETURN_AFTER(ret, type, expr) { \
-    type _trace_return_result = (expr); \
-    _FUNC_RETURN \
-    return _trace_return_result; \
-}
-#define _FUNC_RETURN_VOID_AFTER(ret, type, expr) { \
-    (expr); \
-    _FUNC_RETURN \
-    return; \
-}
-#define _TRACE_CLOSE_AFTER(ret, type, expr) { \
-    type _trace_return_result = (expr); \
-    _TRACE_CLOSE \
-    return _trace_return_result; \
-}
-#define _TRACE_CLOSE_VOID_AFTER(ret, type, expr) { \
-    (expr); \
-    _TRACE_CLOSE \
-    return; \
-}
-#define _FUNC_RETURN_TRACE_CLOSE_AFTER(ret, type, expr) { \
-    type _trace_return_result = (expr); \
-    _FUNC_RETURN \
-    _TRACE_CLOSE \
-    return _trace_return_result; \
-}
-#define _FUNC_RETURN_TRACE_CLOSE_VOID_AFTER(ret, type, expr) { \
-    (expr); \
-    _FUNC_RETURN \
-    _TRACE_CLOSE \
-    return; \
-}
-
 /*
  * Macros used in the instrumentation.
- * 2 versions: _TRACE_MODE and _RETRACE_MODE
+ * versions: _TRACE_MODE and _RETRACE_MODE
+ * and similar
  */
 
 #if defined _TRACE_MODE && defined _RETRACE_MODE
@@ -654,7 +80,7 @@ extern bool _trace_pointer_call;
 
 #define _TRACE_OPEN(fname)  _IS_RETRACE( ,_trace_open((fname))); \
                             _IS_RETRACE( ,_trace_buf_pos = 0); \
-                            _IS_RETRACE( ,_trace_ie_byte = _TRACE_IE_BYTE_INIT); \
+                            _IS_RETRACE( ,_trace_ie_byte = _TRACE_SET_IE_INIT); \
                             _TRACE_POINTER_CALL_SET;
 #define _TRACE_CLOSE        _IS_RETRACE(_RETRACE_END() ,_trace_close())
 
@@ -690,7 +116,7 @@ extern bool _trace_pointer_call;
 
 #define _TRACE_OPEN(fname)  _trace_open((fname)); \
                             _trace_buf_pos = 0; \
-                            _trace_ie_byte = _TRACE_IE_BYTE_INIT; \
+                            _trace_ie_byte = _TRACE_SET_IE_INIT; \
                             _TRACE_POINTER_CALL_SET;
 #define _TRACE_CLOSE        ;_trace_close();
 
@@ -780,54 +206,6 @@ extern bool _trace_pointer_call;
 
 #elif _CBMC_MODE
 
-#include "retrace.h"
-
-#define _RETRACE_CBMC(l, n) { \
-    __CPROVER_assume(retrace_i < retrace_arr_len); \
-    __CPROVER_assume(retrace_arr[retrace_i].letter == l); \
-    __CPROVER_assume(retrace_arr[retrace_i].num == n); \
-    retrace_i += 1; \
-}
-
-#define _RETRACE_CASE_CBMC(num, bit_cnt) ; \
-    for (int i = bit_cnt-1; i >= 0; i--) { \
-        if (num & (1 << i)) { \
-            _RETRACE_CBMC('I', 0); \
-        } else { \
-            _RETRACE_CBMC('O', 0); \
-        } \
-    }
-
-#define _RETRACE_END_CBMC() { \
-    _RETRACE_CBMC('E', 0); \
-    __CPROVER_assume(retrace_i == retrace_arr_len); \
-    /* now check all past assertions */ \
-    for (int i = 0; i < _retrace_assert_idx; i++) { \
-        /* does not work: \
-        __CPROVER_assert(_retrace_asserts[i], \
-                         _retrace_assert_names[i]); \
-         * alternative: */ \
-        assert(_retrace_asserts[i]); \
-    } \
-}
-
-#define _RETRACE_SETJMP_CBMC(setjmp_stmt) ({ \
-    _RETRACE_CBMC('T', 0); \
-    _trace_setjmp_idx ++; \
-    int cur_setjmp_idx = _trace_setjmp_idx; \
-    int setjmp_res = setjmp_stmt; \
-    if (setjmp_res != 0) { \
-        _RETRACE_CBMC('J', _trace_setjmp_idx - cur_setjmp_idx); \
-    } \
-    setjmp_res; \
-})
-
-#define _RETRACE_FUNC_CBMC(num) \
-    if (_TRACE_CALL_CHECK) { \
-        _RETRACE_CBMC('C', num); \
-        _TRACE_POINTER_CALL_RESET; \
-    }
-
 #define _IF                 ;_RETRACE_CBMC('I', 0);
 #define _ELSE               ;_RETRACE_CBMC('O', 0);
 #define _CONDITION(cond)    cond
@@ -892,8 +270,10 @@ extern bool _trace_pointer_call;
 
 #endif // _TRACE_MODE or _RETRACE_MODE
 
+
+// quickfix...
 #ifdef BOOL_ALREADY_DEFINED
-#undef bool
+    #undef bool
 #endif
 
 #ifdef __cplusplus
