+++ include/src_tracer/trace_buf.h
@@ -1,0 +1,22 @@
+/*
+   src_tracer/trace_buf.h
+   anything related to the trace buffer
+*/
+#ifndef SRC_TRACER_TRACE_BUF_H
+#define SRC_TRACER_TRACE_BUF_H
+
+#include "src_tracer/constants.h"
+
+#ifndef BYTE_TRACE
+    extern unsigned char _trace_ie_byte;
+#endif
+
+extern unsigned char _trace_buf[TRACE_BUF_SIZE];
+extern int _trace_buf_pos;
+
+extern void _trace_write(const void *buf);
+#ifndef EFFICIENT_TEXT_TRACE
+    extern void _trace_write_text(const void *buf, unsigned long count);
+#endif
+
+#endif //SRC_TRACER_TRACE_BUF_H
